+++ collect_app/src/androidTest/java/org/odk/collect/android/storage/StorageMigrationCompletedBannerTest.java
@@ -1,20 +1,52 @@
 package org.odk.collect.android.storage;
 
+import android.Manifest;
+
 import androidx.test.rule.ActivityTestRule;
+import androidx.test.rule.GrantPermissionRule;
 
 import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.RuleChain;
 import org.odk.collect.android.activities.MainMenuActivity;
-import org.odk.collect.android.support.StorageMigrationCompletedRule;
+import org.odk.collect.android.injection.config.AppDependencyModule;
+import org.odk.collect.android.storage.migration.StorageMigrationRepository;
+import org.odk.collect.android.storage.migration.StorageMigrationResult;
+import org.odk.collect.android.support.ResetStateRule;
 import org.odk.collect.android.support.pages.MainMenuPage;
+
+import javax.inject.Singleton;
+
+import dagger.Provides;
+
+import static org.odk.collect.android.support.CollectHelpers.overrideAppDependencyModule;
 
 public class StorageMigrationCompletedBannerTest {
 
     @Rule
-    public ActivityTestRule<MainMenuActivity> main = new ActivityTestRule<>(MainMenuActivity.class);
+    public RuleChain copyFormChain = RuleChain
+            .outerRule(GrantPermissionRule.grant(
+                    Manifest.permission.READ_EXTERNAL_STORAGE,
+                    Manifest.permission.WRITE_EXTERNAL_STORAGE
+            ))
+            .around(new ResetStateRule());
 
     @Rule
-    public StorageMigrationCompletedRule storageMigrationCompletedRule = new StorageMigrationCompletedRule();
+    public ActivityTestRule<MainMenuActivity> main = new ActivityTestRule<MainMenuActivity>(MainMenuActivity.class) {
+        @Override
+        protected void beforeActivityLaunched() {
+            overrideAppDependencyModule(new AppDependencyModule() {
+                @Provides
+                @Singleton
+                public StorageMigrationRepository providesStorageMigrationRepository() {
+                    StorageMigrationRepository storageMigrationRepository = new StorageMigrationRepository();
+                    storageMigrationRepository.setResult(StorageMigrationResult.SUCCESS);
+                    return storageMigrationRepository;
+                }
+            });
+            new StorageStateProvider().enableUsingScopedStorage();
+        }
+    };
 
     @Test
     public void when_storageMigrationCompleted_should_bannerBeVisibleAndPersistScreenRotation() {
@@ -42,7 +74,7 @@ public class StorageMigrationCompletedBannerTest {
                 .assertStorageMigrationCompletedBannerIsNotDisplayed()
                 .rotateToPortrait(new MainMenuPage(main))
                 .assertStorageMigrationCompletedBannerIsNotDisplayed()
-                .reopenApp()
+                .recreateActivity()
                 .assertStorageMigrationCompletedBannerIsNotDisplayed();
     }
 
@@ -50,7 +82,7 @@ public class StorageMigrationCompletedBannerTest {
     public void when_storageMigrationCompleted_should_bannerBeVisibleAndDismissAfterReopeningApp() {
         new MainMenuPage(main)
                 .assertStorageMigrationCompletedBannerIsDisplayed()
-                .reopenApp()
+                .recreateActivity()
                 .assertStorageMigrationCompletedBannerIsNotDisplayed();
     }
 }
